homework 10_1 done
