+++ README.md
@@ -1,40 +1,14 @@
-# Проект "Банк МММ"
+# My Bank Operations Project
 
-## Описание:
+## Описание проекта
+Проект предназначен для работы с банковскими операциями. 
+В частности, содержит функции для:
+- фильтрации операций по статусу (`filter_by_state`),
+- сортировки операций по дате (`sort_by_date`).
 
-Проект "Банк МММ" - это веб-приложение на Python для создания и управления персональным банком. Пользователи могут создавать данные своих карт, редактировать их, а мы можем снимать с их карт деньги!
+Проект написан на Python и придерживается [GitFlow](https://datasithu.medium.com/a-gentle-introduction-to-git-flow-7f849c9b7f2) для ведения разработки.
 
-## Установка:
-
-1. Клонируйте репозиторий:
-```
-git clone https://github.com/username/Esp47291
-```
-
-2. Установите зависимости:
-```
-pip install -r requirements.txt
-```
-
-3. Создайте базу данных и выполните миграции:
-```
-python manage.py migrate
-```
-
-4. Запустите локальный сервер:
-```
-python manage.py runserver
-```
-## Использование:
-
-1. Перейдите на страницу в вашем веб-браузере.
-2. Создайте новую учетную запись или войдите существующей.
-3. Ждите новую запись в банке и снимите деньги
-
-## Документация:
-
-Дополнительную информацию о структуре проекта и API можно найти в [документации](docs/README.md).
-
-## Лицензия:
-
-Проект распространяется под [лицензией MIT](LICENSE).
+## Установка
+1. Склонируйте репозиторий:
+   ```bash
+   git clone https://github.com/username/my_bank_operations_project.git
