+++ tests/test_main.py
@@ -1,5 +1,6 @@
 def test_one():
     assert 1 == 1
 
+
 def new_changes():
     pass
